add related names on user_roles and author / add functionnal serializer for full user object
